+++ app/Http/Livewire/Galaxy/Events/Show.php
@@ -27,11 +27,18 @@ class Show extends Component
 
     public function getPagesProperty()
     {
-        return [
-            ['value' => 'details', 'label' => 'Dashboard', 'href' => route('galaxy.events.show', ['event' => $this->event, 'page' => 'dashboard']), 'icon' => 'heroicon-o-home', 'active' => $this->page === 'dashboard'],
-            ['value' => 'reservations', 'label' => 'Reservations', 'href' => route('galaxy.events.show', ['event' => $this->event, 'page' => 'reservations']), 'icon' => 'heroicon-o-camera', 'active' => $this->page === 'reservations'],
-            ['value' => 'orders', 'label' => 'Orders', 'href' => route('galaxy.events.show', ['event' => $this->event, 'page' => 'orders']), 'icon' => 'heroicon-o-ticket', 'active' => $this->page === 'orders'],
-            ['value' => 'workshops', 'label' => 'Workshops', 'href' => route('galaxy.events.show', ['event' => $this->event, 'page' => 'workshops']), 'icon' => 'heroicon-o-ticket', 'active' => $this->page === 'workshops'],
-        ];
+        $pages = [['value' => 'details', 'label' => 'Dashboard', 'href' => route('galaxy.events.show', ['event' => $this->event, 'page' => 'dashboard']), 'icon' => 'heroicon-o-home', 'active' => $this->page === 'dashboard']];
+
+        if($this->event->settings->reservations) {
+            $pages[] = ['value' => 'reservations', 'label' => 'Reservations', 'href' => route('galaxy.events.show', ['event' => $this->event, 'page' => 'reservations']), 'icon' => 'heroicon-o-camera', 'active' => $this->page === 'reservations'];
+        }
+
+        $pages[] = ['value' => 'orders', 'label' => 'Orders', 'href' => route('galaxy.events.show', ['event' => $this->event, 'page' => 'orders']), 'icon' => 'heroicon-o-ticket', 'active' => $this->page === 'orders'];
+
+        if($this->event->settings->has_workshops) {
+            $pages[] = ['value' => 'workshops', 'label' => 'Workshops', 'href' => route('galaxy.events.show', ['event' => $this->event, 'page' => 'workshops']), 'icon' => 'heroicon-o-ticket', 'active' => $this->page === 'workshops'];
+        }
+
+        return $pages;
     }
 }