+++ resources/views/layouts/galaxy.blade.php
@@ -11,14 +11,16 @@
     <!-- Styles -->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/flatpickr/dist/flatpickr.min.css">
     <link rel="stylesheet" type="text/css" href="https://unpkg.com/trix@1.2.3/dist/trix.css">
+    <link rel="stylesheet" href="https://unpkg.com/easymde/dist/easymde.min.css">
     <link rel="stylesheet" href="{{ mix('css/app.css') }}">
     <livewire:styles />
     @stack('styles')
 
     <!-- Scripts -->
+    <livewire:scripts />
     <script src="{{ asset('js/app.js') }}" defer></script>
     <script src="https://unpkg.com/trix@1.2.3/dist/trix.js" defer></script>
-    <livewire:scripts />
+    <script src="https://unpkg.com/easymde/dist/easymde.min.js"></script>
     @stack('scripts')
 </head>